\name{extra2010}
     \docType{data}
\alias{extra2010}
\title{World Development Indicators Extra Variables}
\description{Extra variables to merge with indicator dataframes}
\author{World Bank WDI}
     \usage{extra2010}
     \format{A dataframe}
     \source{World Bank (2010)}
     \keyword{datasets}
