% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WDI.R
\name{languages_supported}
\alias{languages_supported}
\title{List of supported languages}
\usage{
languages_supported()
}
\value{
A list of fully and partially supported languages.
}
\description{
This prints two lists of languages, the fully supported ones and the locally supported ones:
\itemize{
\item the languages in the category "fully" will return translated names and other info for all countries.
\item the languages in the category "partially" will return translated names and other info only for the country they represent.
}
}
\details{
For example, choosing "vi" (for Vietnamese) will translate "Vietnam" in the dataset but other country names won't be translated and will be empty.
}
