% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WDI-package.R
\docType{package}
\name{WDI-package}
\alias{WDI-package}
\title{Search, extract, and format data from the World Bank's World Development Indicators}
\description{
Search, extract and format data from the World Bank's World Development Indicators
}
\details{
The \code{WDI} function provides convenient access to the World Bank's WDI
download API.  For fast searching, the WDI package ships with a local list of
available data series. This local list can be updated to the latest version
using the \code{WDIcache} function.
}
\references{
\url{http://www.umich.edu/~varel}
    \url{https://github.com/vincentarelbundock/WDI}
}
\author{
Vincent Arel-Bundock \email{varel@umich.edu}
}
\keyword{package}
