% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WDI.R
\name{wdi.query}
\alias{wdi.query}
\title{Internal function to build API call}
\usage{
wdi.query(
  indicator = "NY.GDP.PCAP.CD",
  country = "all",
  start = NULL,
  end = NULL,
  latest = NULL,
  language = "en"
)
}
\description{
Internal function to build API call
}
\keyword{internal}
