% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WDI.R
\name{WDIbulk}
\alias{WDIbulk}
\title{Download all the WDI indicators at once.}
\usage{
WDIbulk(timeout = 600)
}
\arguments{
\item{timeout}{integer maximum number of seconds to wait for download}
}
\value{
Data frame

a list of 6 data frames: Data, Country, Series, Country-Series,
Series-Time, FootNote
}
\description{
Download all the WDI indicators at once.
}
\author{
Vincent Arel-Bundock \email{vincent.arel-bundock@umontreal.ca}
}
