\name{WDIseries}
\alias{WDIseries}
\title{World Development Indicators series...}
\usage{series}
\description{World Development Indicators series}
\details{Code, name, description, source, and topic of all WDI series. Retrieved 2012-03-31

\itemize{
\item indicator code  
\item name name of the series
\item topic topic/category of the variable
\item description detailed description
\item source organization which produced the variable 
}}
\docType{data}
\keyword{datasets}
