% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WDI-package.R
\docType{package}
\name{WDI-package}
\alias{WDI-package}
\title{Search, extract, and format data from the World Bank}
\description{
Search, extract and format data from the World Bank's World Development Indicators
}
\details{
The \code{WDI} function provides convenient access to over 40 databases
hosted by the World Bank, including the World Development Indicators (WDI),
International Debt Statistics, Doing Business, Human Capital Index, and
Sub-national poverty indicators.  For fast searching, the WDI package ships
with a local list of available data series. This local list can be updated
to the latest version using the \code{WDIcache} function.
}
\references{
\url{http://arelbundock.com}
    \url{https://github.com/vincentarelbundock/WDI}
}
\author{
Vincent Arel-Bundock \email{vincent.arel-bundock@umontreal.ca}
}
\keyword{package}
