\name{WDIsearch}
\alias{WDIsearch}
\title{Search names and descriptions of available WDI series...}
\usage{WDIsearch(string="gdp", field="name", short=TRUE, cache)}
\description{Search names and descriptions of available WDI series}
\details{Data frame with series code, name, description, and source for the WDI series which match the given criteria}
\value{Data frame with code, name, source, and description of all series which match the criteria.}
\author{Vincent Arel-Bundock \email{varel@umich.edu}}
\note{Downloading all series information from the World Bank website can take time.
The \code{WDI} package ships with a local data object with information on all the series
available on 2012-03-31. You can update this database by retrieving a new list using \code{WDIcache}, and  then
feeding the resulting object to \code{WDI} via the \code{cache} argument.}
\arguments{\item{string}{Character string. Search for this string.}
\item{field}{Character string. Search this field. Admissible fields: 'indicator', 'name', 'topic', 'description', 'source'}
\item{short}{TRUE: Returns only the indicator's code and name. FALSE: Returns the indicator's code, name, topic, description, and source.}
\item{cache}{Data frame generated by the \code{WDIcache} function. If omitted, \code{WDIsearch} will search a local list of series updated on 2012-03-31.}}
\examples{WDIsearch(string='gdp', field='name', cache=NULL)
WDIsearch(string='AG.AGR.TRAC.NO', field='indicator', cache=NULL)}
